requirements.txt
sentence-transformers 
faiss-cpu 
beautifulsoup4 
nltk 
numpy
pydantic 
sec_edgar_downloader
bs4
nltk
sentence_transformers